refactor: split e2e tests by domain
Split the monolith e2e_test.go into domain-specific files:
- harness_test.go — test harness, helpers, emptyWidget
- sidebar_test.go — sidebar tab click, toggle, switching, width, overflow, focus
- panel_test.go — bottom panel tab click, remove, toggle
- explorer_test.go — key nav, expand/collapse, open file, click, scroll
- editor_test.go — startup, menu bar, new file, palette, find, theme
- changes_test.go — changes nav, refresh
